# docs/sacp/openapi-agent.yaml
openapi: 3.0.3
info:
  title: WSIM Agent Commerce API
  description: |
    WSIM Agent Commerce API - Enables AI agents to make authenticated purchases on behalf of users.

    ## Overview

    This API supports the SimToolBox Agent Commerce Protocol (SACP), allowing AI agents to:
    - Register with user wallets via OAuth Device Authorization or pairing codes
    - Authenticate using OAuth 2.0 client credentials
    - Request payment tokens for purchases
    - Handle step-up authorization for transactions exceeding limits

    ## Authorization Flows

    ### Flow 1: Device Authorization Grant (RFC 8628) - RECOMMENDED
    Agent-initiated flow for autonomous agents:
    1. Agent calls `POST /api/agent/v1/oauth/device_authorization`
    2. Receives `device_code` and `user_code` (format: WSIM-XXXXXX-XXXXXX)
    3. Agent displays `user_code` and `verification_uri` to user
    4. User enters code in WSIM mobile app
    5. Agent polls `POST /api/agent/v1/oauth/token` with `grant_type=urn:ietf:params:oauth:grant-type:device_code`
    6. On approval, receives `client_id` and `client_secret`
    7. Uses client credentials to obtain access tokens

    ### Flow 2: Pairing Code (User-Initiated)
    User-initiated flow when user has pre-generated code:
    1. User generates pairing code in mobile app
    2. User provides code to agent (WSIM-XXXXXX-XXXXXX format)
    3. Agent calls `POST /api/agent/v1/access-request` with the code
    4. Agent polls until user approves
...

...

...

...

...

...

...

...
  contact:
...
servers:
...
paths:
...
  /api/agent/v1/oauth/device_authorization:
    post:
      tags:
        - OAuth
      summary: Start device authorization
      description: |
        OAuth 2.0 Device Authorization Grant (RFC 8628).

        **This is the recommended flow for AI agents.** The agent initiates authorization
        and receives a user code to display to the user.

        **Flow:**
        1. Agent calls this endpoint with agent details
        2. Receives `device_code`, `user_code`, and `verification_uri`
        3. Agent displays `user_code` to user (format: WSIM-XXXXXX-XXXXXX)
        4. User enters code in WSIM mobile app (or visits `verification_uri_complete`)
        5. Agent polls `/oauth/token` with `grant_type=urn:ietf:params:oauth:grant-type:device_code`
        6. On approval, receives client credentials

        **Expiration**: Device codes expire in 15 minutes.
        **Polling**: Poll every 5 seconds (as indicated by `interval` field).
      operationId: startDeviceAuthorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agent_name
              properties:
                agent_name:
                  type: string
                  description: Display name for the agent (shown to user)
                  maxLength: 100
                  example: "Shopping Assistant"
                agent_description:
                  type: string
                  description: Description of what the agent does
                  maxLength: 500
                  example: "Helps you find and purchase products online"
                scope:
                  type: string
                  description: Space-separated list of requested permissions
                  example: "browse cart purchase"
                  default: "browse"
                spending_limits:
                  type: object
                  description: Requested spending limits (user may lower these)
                  properties:
                    per_transaction:
                      type: number
                      description: Max per-transaction limit
                      default: 50
                    daily:
                      type: number
                      description: Daily spending limit
                      default: 200
                    monthly:
                      type: number
                      description: Monthly spending limit
                      default: 1000
                    currency:
                      type: string
                      default: "CAD"
      responses:
        '200':
          description: Device authorization initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceAuthorizationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthErrorResponse'
  /api/agent/v1/oauth/token:
    post:
      tags:
        - OAuth
      summary: Exchange credentials for access token
      description: |
        OAuth 2.0 token endpoint supporting multiple grant types:

        **1. Client Credentials Grant** (`client_credentials`)
        Exchange client credentials for an access token:
        ```
        grant_type=client_credentials
        client_id=agent_xxx
        client_secret=sk_agent_xxx
        ```

        **2. Device Authorization Grant** (`urn:ietf:params:oauth:grant-type:device_code`)
        Poll for device authorization approval:
        ```
        grant_type=urn:ietf:params:oauth:grant-type:device_code
        device_code=<device_code from device_authorization>
        ```

        For device_code grant, possible responses:
        - `authorization_pending`: User hasn't approved yet, keep polling
        - `access_denied`: User rejected the request
        - `expired_token`: Authorization expired
        - Success: Returns client credentials (not access token)
      operationId: getAgentToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - grant_type
              properties:
                grant_type:
                  type: string
                  enum:
                    - client_credentials
                    - urn:ietf:params:oauth:grant-type:device_code
                  description: Grant type
                client_id:
                  type: string
                  description: Agent client ID (required for client_credentials)
                  example: agent_abc123def456
                client_secret:
                  type: string
                  description: Agent client secret (required for client_credentials)
                  example: sk_agent_xxxxxxxxxxxxxxxxxxxx
                device_code:
                  type: string
                  description: Device code (required for device_code grant)
                scope:
                  type: string
                  description: Optional scope restriction
                  example: "browse cart purchase"
      responses:
        '200':
          description: Token issued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthErrorResponse'
        '403':
          description: Agent suspended or delegation expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthErrorResponse'
  /api/agent/v1/oauth/introspect:
    post:
      tags:
        - OAuth
      summary: Introspect agent token
      description: |
        Token introspection endpoint for merchants (SSIM) to validate agent tokens.

        Returns full agent context including:
        - Agent identity and permissions
        - Owner information
        - Spending limits and current usage
        - Token validity and expiration

        **Caching**: Results may be cached for up to 60 seconds (see Q18).
      operationId: introspectAgentToken
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: The agent access token to introspect
      responses:
        '200':
          description: Token introspection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntrospectionResponse'
        '401':
          description: Invalid introspection credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/agent/v1/oauth/revoke:
    post:
      tags:
        - OAuth
      summary: Revoke agent token
      description: |
        Revoke an agent's access token. Use when:
        - User revokes agent access
        - Agent is compromised
        - Session should be terminated
      operationId: revokeAgentToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: The agent access token to revoke
                client_id:
                  type: string
                  description: Agent client ID
                client_secret:
                  type: string
                  description: Agent client secret
      responses:
        '200':
          description: Token revoked successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthErrorResponse'

  # =============================================================================
  # Payment Endpoints
  # =============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Mobile app JWT authentication

    agentBearerAuth:
      type: http
      scheme: bearer
      description: Agent OAuth access token

    basicAuth:
      type: http
      scheme: basic
      description: Basic auth for introspection (client_id:client_secret)
  schemas:
...
    DeviceAuthorizationResponse:
      type: object
      required:
        - device_code
        - user_code
        - verification_uri
        - expires_in
        - interval
      properties:
        device_code:
          type: string
          description: Code for polling token endpoint
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_code:
          type: string
          description: Code for user to enter (format WSIM-XXXXXX-XXXXXX)
          example: "WSIM-ABC123-DEF456"
        verification_uri:
          type: string
          format: uri
          description: URL for user to visit to enter code
          example: "https://wsim.banksim.ca/m/device"
        verification_uri_complete:
          type: string
          format: uri
          description: URL with code pre-filled
          example: "https://wsim.banksim.ca/m/device?code=WSIM-ABC123-DEF456"
        expires_in:
          type: integer
          description: Code lifetime in seconds (900 = 15 minutes)
          example: 900
        interval:
          type: integer
          description: Minimum polling interval in seconds
          example: 5
...
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: The agent access token
          example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          enum: [Bearer]
          description: Token type (always "Bearer")
        expires_in:
          type: integer
          description: Token lifetime in seconds
          example: 3600
        scope:
          type: string
          description: Granted scopes
          example: "browse cart purchase"
    IntrospectionResponse:
      type: object
      required:
        - active
      properties:
        active:
          type: boolean
          description: Whether the token is valid and active
        client_id:
          type: string
          description: Agent client ID
          example: agent_abc123def456
        agent_id:
          type: string
          description: Internal agent UUID
          example: "550e8400-e29b-41d4-a716-446655440000"
        owner_id:
          type: string
          description: WSIM user ID of agent owner (UUID)
          example: "660e8400-e29b-41d4-a716-446655440001"
        agent_name:
          type: string
          description: Human-readable agent name
          example: "My Shopping Assistant"
        permissions:
          type: array
          items:
            type: string
            enum: [browse, cart, purchase, history]
          description: Agent permissions
          example: ["browse", "cart", "purchase"]
        spending_limits:
          $ref: '#/components/schemas/SpendingLimits'
        spending_usage:
          $ref: '#/components/schemas/SpendingUsage'
        exp:
          type: integer
          description: Token expiration (Unix timestamp)
          example: 1705858800
        iat:
          type: integer
          description: Token issued at (Unix timestamp)
          example: 1705855200
    OAuthErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - access_denied
            - delegation_expired
            - authorization_pending
            - slow_down
            - expired_token
          description: |
            OAuth 2.0 error code:
            - `access_denied`: Agent is suspended/revoked or user denied request
            - `invalid_client`: Invalid client credentials
            - `unsupported_grant_type`: Invalid grant type
            - `authorization_pending`: Device code not yet approved (keep polling)
            - `slow_down`: Polling too fast, increase interval
            - `expired_token`: Device code has expired
        error_description:
          type: string
          description: Human-readable error description
          example: "Agent is suspended"

    # =========================================================================
    # Payment Schemas
    # =========================================================================
...
    SpendingLimits:
      type: object
      required:
        - per_transaction
        - daily
        - monthly
        - currency
      properties:
        per_transaction:
          type: number
          format: decimal
          example: 50.00
        daily:
          type: number
          format: decimal
          example: 200.00
        monthly:
          type: number
          format: decimal
          example: 500.00
        currency:
          type: string
          example: "CAD"
    SpendingUsage:
      type: object
      properties:
        today:
          type: number
          format: decimal
          description: Amount spent today (resets at midnight EST/EDT)
          example: 45.67
        this_month:
          type: number
          format: decimal
          description: Amount spent this month
          example: 245.50
        last_transaction:
          type: string
          format: date-time
          description: Timestamp of last transaction
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "AGENT_NOT_FOUND"
            message:
              type: string
              description: Human-readable error message
              example: "Agent not found"
            details:
              type: object
              description: Additional error context
...